requirements.txt
streamlit==1.32.2
tensorflow==2.18.0
numpy==1.24.3
Pillow==10.0.0
transformers==4.21.0
protobuf==3.20.3
fastapi==0.110.0
uvicorn==0.29.0